[internal] Refactor PyO3 FFI code
# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]
