+++ src/rust/engine/engine_pyo3/src/externs/interface/nailgun.rs
@@ -1,11 +1,12 @@
 // Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-use super::PyExecutor;
 use pyo3::create_exception;
 use pyo3::exceptions::{PyBrokenPipeError, PyException, PyKeyboardInterrupt};
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
+
+use crate::externs::interface::scheduler::PyExecutor;
 use task_executor::Executor;
 
 pub(crate) fn register(py: Python, m: &PyModule) -> PyResult<()> {
